non blocking for put handler
