+++ crates/ursa-index-provider/Cargo.toml
@@ -36,8 +36,8 @@ ursa-store = { path = "../ursa-store" }
 ursa-utils ={ path = "../ursa-utils" }
 ursa-network = { path = "../ursa-network" }
 futures = "0.3.21"
+simple_logger = "4.0.0"
 
 [dev-dependencies]
-simple_logger = "2.1.0"
 ursa-network = { path = "../ursa-network" }
