+++ crates/ursa-index-provider/src/lib.rs
@@ -3,3 +3,4 @@ pub mod config;
 pub mod engine;
 pub mod provider;
 pub mod signed_head;
+mod tests;
